Corrected StreamHeader struct incorreclty using long in the frame sub-struct

+++ Avisynth3/trunk/avisynth/src/import/avi/streamheader.h
@@ -56,10 +56,10 @@ struct StreamHeader
   
   struct Rectange
   {
-    long left;
-    long top;
-    long right;
-    long bottom;
+    short left;
+    short top;
+    short right;
+    short bottom;
   } frame;
